feat(core): add function to add new colors (#19)
# Proposed changes

Introduced logic to add new color slices on both widget' state and
library. It also sync the library if needed (ex. if the color on library
does not match the one on the state / if no related color is found)

wip #15

## Types of changes

- [x] ✨ New feature (non-breaking change which adds functionality or
enhancements)
- [ ] 🐛 Bugfix (non-breaking change which fixes an issue)
- [ ] 💥 Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] ✅ Tests (added tests for an existing feature)
- [ ] 📚 Chore / Documentation Update (if none of the other choices
apply)
- [ ] 🙌 Other (please, write a clear and concise description of the
proposal in the section above)

## Checklist

- [x] :scissors: I have ensured that the PR is concise and broken down
if needed
- [x] 👀 I have read the [README](../README.md) doc
- [x] 🧪 I have added tests that prove my fix is effective or that my
feature works
- [ ] 📚 I have added necessary documentation (if appropriate)
- [x] 🔀 Any dependent changes have been merged and published in
downstream modules